added the mocked login service in the tests

+++ app/src/main/java/ch/epfl/sdp/appart/login/FirebaseLoginService.java
@@ -44,6 +44,7 @@ public class FirebaseLoginService implements LoginService {
 
     @Override
     public CompletableFuture<User> loginWithEmail(String email, String password) {
+        if (getCurrentUser() == null) throw new IllegalStateException("current user must not already be set when authentication");
         return handleEmailAndPasswordMethod(email, password,
                 mAuth.signInWithEmailAndPassword(email, password));
     }
@@ -69,6 +70,7 @@ public class FirebaseLoginService implements LoginService {
 
     @Override
     public CompletableFuture<User> createUser(String email, String password) {
+        if (getCurrentUser() != null) throw new IllegalStateException("the current user cannot be set");
         return handleEmailAndPasswordMethod(email, password,
                 mAuth.createUserWithEmailAndPassword(email, password));
     }
@@ -76,6 +78,7 @@ public class FirebaseLoginService implements LoginService {
     @Override
     public CompletableFuture<Void> updateEmailAddress(String email) {
         if (email == null) throw new IllegalArgumentException();
+        if (getCurrentUser() == null) throw new IllegalStateException("current user must be set when updating the email");
         return setUpFuture(getCurrentFirebaseUser().updateEmail(email)
                 , result -> result);
     }
@@ -83,18 +86,21 @@ public class FirebaseLoginService implements LoginService {
     @Override
     public CompletableFuture<Void> updatePassword(String password) {
         if (password == null) throw new IllegalArgumentException();
+        if (getCurrentUser() == null) throw new IllegalStateException("current user must be set when updating the password");
         return setUpFuture(getCurrentFirebaseUser().updatePassword(password),
                 result -> result);
     }
 
     @Override
     public CompletableFuture<Void> sendEmailVerification() {
+        if (getCurrentUser() == null) throw new IllegalStateException("current user must be set when sending verification mail");
         return setUpFuture(getCurrentFirebaseUser().sendEmailVerification(),
                 result -> result);
     }
 
     @Override
     public CompletableFuture<Void> deleteUser() {
+        if (getCurrentUser() == null) throw new IllegalStateException("current user must be set when deleting it");
         return setUpFuture(getCurrentFirebaseUser().delete(),
                 result -> result);
     }
@@ -102,6 +108,7 @@ public class FirebaseLoginService implements LoginService {
     @Override
     public CompletableFuture<Void> reAuthenticateUser(String email, String password) {
         if (email == null || password == null) throw new IllegalArgumentException();
+        if (getCurrentUser() == null) throw new IllegalStateException("current user must be set when reAuthentication");
         return setUpFuture(getCurrentFirebaseUser().reauthenticate(
                 EmailAuthProvider.getCredential(email, password)),
                 result -> result);
@@ -115,7 +122,7 @@ public class FirebaseLoginService implements LoginService {
 
     /**
      * @return the current Firebase user
-     * @throws IllegalArgumentException if no user is set
+     * @throws IllegalStateException if no user is set
      */
     private FirebaseUser getCurrentFirebaseUser() {
         FirebaseUser user = this.mAuth.getCurrentUser();
